Fix TypeError in playRound function
- The toLowerCase method on the passed argument was removed
in the playRound function. This is because it is redundant now
that the game does not rely on text input from the user.

- The event listener that initialises the game once the DOM
is loaded was refactored to no longer require an anonymous function.
Instead, it simply passes a reference to the required function.

- The getComputerChoice function was further refined to calculate
a random value based on the length of the array, as opposed to a set
value. This means that further choices can be added easier, if needed.